music-map-phase1/backend/requirements.txt
Flask
Flask-Session
spotipy
python-dotenv
musicbrainzngs
geopy
wikipedia-api
requests
beautifulsoup4

